requirements.txt
streamlit
openai==0.28
pandas
pdfplumber
python-dotenv
openpyxl
requests
Pillow
PyMuPDF
